requirements.txt
pdfplumber~=0.11.5
pyzbar~=0.1.9
numpy~=2.2.3
pillow~=11.1.0
pip~=24.3.1
pathlib~=1.0.1
configparser~=7.1.0
attrs~=24.2.0
wheel~=0.43.0
cryptography~=43.0.3
filelock~=3.17.0
setuptools~=75.5.0
pytest~=8.1.1
pytz~=2025.1
cffi~=1.17.1
scipy~=1.15.2
platformdirs~=4.2.0
spire~=0.4.2
pandas~=2.2.3
packaging~=24.0
pdf2image~=1.17.0
pycparser~=2.22
pypdfium2~=4.30.1
pytesseract~=0.3.13
bcrypt~=4.2.0
fitz~=0.0.1.dev2
